Refactor file management backend and frontend
- Updated `FileController` to use `fileManifestId` in the `DownloadFile` endpoint and improved response messaging.
- Replaced old migrations with a new consolidated migration (`20251014051137_Initial`) to introduce `file_manifests` and `file_manifest_chunks` tables, replacing `blobs` and `blob_chunks`.
- Added `size_bytes` as a `bigint` column in the `chunks` table and updated relationships in `CottonDbContextModelSnapshot`.
- Enhanced the frontend API layer in `files.ts` with new interfaces (`FileManifestDto`, `CreateFileRequest`) and functions (`listFiles`, `getDownloadUrl`, `createFileFromChunks`).
- Refactored `FilesPage.tsx` to display a grid of file cards, integrate with the new backend, and improve the file upload flow.
- Updated `config.ts` to include a new `files` API endpoint and a utility function `buildDownloadUrl`.
- Improved code readability, error handling, and UI design for better maintainability and user experience.

+++ Sources/Cotton.Server/Migrations/CottonDbContextModelSnapshot.cs
@@ -31,8 +31,8 @@ namespace Cotton.Server.Migrations
                         .HasColumnType("bytea")
                         .HasColumnName("sha256");
 
-                    b.Property<int>("SizeBytes")
-                        .HasColumnType("integer")
+                    b.Property<long>("SizeBytes")
+                        .HasColumnType("bigint")
                         .HasColumnName("size_bytes");
 
                     b.HasKey("Sha256");
